[device]: ag5648 device drivers           1. Use the common driver dni_emc2305.c
          ag9032v1 device drivers
          1. Move dni_emc2305.c to be the common driver
          2. Remove at24.c
