+++ platform/broadcom/sonic-platform-modules-delta/ag9032v1/modules/Makefile
@@ -1,2 +1,2 @@
-obj-m := at24.o dni_ag9032v1_psu.o dni_emc2305.o delta_ag9032v1_platform.o delta_ag9032v1_cpupld.o
+obj-m := dni_ag9032v1_psu.o dni_emc2305.o delta_ag9032v1_platform.o delta_ag9032v1_cpupld.o
